requirements.txt
openai
python-dotenv
gradio
elevenlabs

